Move Android "api" module to `platform_impl` (#1420)
As seen on the `v0.23_reworks` branch as well, this module doesn't
contain a GL API but the platform-specific implementation for Android,
built on top of EGL.  It should hence reside in `platform_impl`, which
is currently a public reexport of every public item inside
`api::android` anyway.
