+++ glutin/src/platform/android.rs
@@ -1,4 +1,4 @@
-#![cfg(any(target_os = "android"))]
+#![cfg(target_os = "android")]
 
 use crate::platform::ContextTraitExt;
 use crate::{Context, ContextCurrentState};
